added possiblity to use fixed or "inherited" compression level for each file seperately + related fixes.
inhearited means to use the same level used for the solid block and the headers.
inherited levels can be changed with repacking ('lvpak -c# r archive.lvpa')
fixed some problems with the 'packed' flag that was used to indicate if a file should be packed or not.
now it is only used to indicate that a file was packed and must be unpacked, when loading.